feat(arch): :building_construction: Hello world with riverpod
We downgrade the Dart SDK version to disable "sound null safety",
as the package we'll add doesn't support it yet. Then, adds the
flutter_riverpod package. Import it, wrap entire app with
"ProviderScope". Create a view with the provider from riverpod package.
Write tests for it, then you have it!